pyproject.toml
[project]
name = "analytical-platform-ui"
version = "0.2.0"
description = "Analytical Platform UI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib==1.3.2",
    "boto3==1.36.6",
    "django==5.1.9",
    "django-extensions==3.2.3",
    "django-structlog==8.1.0",
    "gunicorn==23.0.0",
    "model-bakery==1.20.1",
    "moto[all]==5.0.18",
    "psycopg[binary]==3.2.9",
    "pyjwt==2.10.1",
    "pytest==8.3.3",
    "pytest-django==4.9.0",
    "python-dotenv==1.0.1",
    "requests==2.32.4",
    "sentry-sdk==2.20.0",
    "uvicorn[standard]==0.32.0",
]


[dependency-groups]
dev = [
    "django-debug-toolbar==4.4.6",
    "pre-commit==4.0.1",
    "pytest-dotenv==0.5.2",
    "ruff==0.12.2",
]


[tool.ruff]
line-length = 100
exclude = [
  "**/migrations/*",
]


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.pytest.ini_options]
addopts = "-v --tb=auto --ds=ap.settings.test"
testpaths = ["tests"]
